pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "toke-models"
version = "0.1.0"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [{name = "M. Karwalski"}]

[project.optional-dependencies]
cuda = [
    "torch>=2.1",
    "transformers>=4.36",
    "peft>=0.7",
    "bitsandbytes>=0.41",
    "datasets>=2.14",
    "accelerate>=0.25",
]
mlx = [
    "mlx>=0.22",
    "mlx-lm>=0.21",
    "pyyaml>=6",
]
dev = [
    "pytest>=7",
    "ruff>=0.4",
    "mypy>=1.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
strict = true
python_version = "3.11"
